Fix code format (pre-commit)

+++ tests/experimental/sgn__test.py
@@ -9,25 +9,27 @@ try:
     from tests.accuracy_utils import gems_assert_close
 except ImportError:
     # Fallback values when running outside pytest
-    
+
     def gems_assert_close(res, ref, dtype, **kwargs):
         # Simple fallback comparison
         torch.testing.assert_close(res, ref, **kwargs)
 
+
 import pytest
+import torch
 import triton
 
 import flag_gems
 from flag_gems.experimental_ops.sgn_ import sgn_ as gems_sgn_
-
-import torch
 
 
 @pytest.mark.sgn_
 @pytest.mark.parametrize("shape", [(2, 3), (128, 256), (512, 512)])
 @pytest.mark.parametrize("dtype", [torch.float32, torch.float16, torch.bfloat16])
 def test_sgn__tensor(shape, dtype):
-    input_tensor = torch.randn(shape, device=flag_gems.device, dtype=torch.float32).to(dtype)
+    input_tensor = torch.randn(shape, device=flag_gems.device, dtype=torch.float32).to(
+        dtype
+    )
     flat = input_tensor.view(-1)
     if flat.numel() >= 1:
         flat[0] = flat.new_zeros(1)
@@ -44,6 +46,7 @@ def test_sgn__tensor(shape, dtype):
 
     gems_assert_close(act_out, ref_out, dtype=dtype)
 
+
 @pytest.mark.sgn_
 @pytest.mark.parametrize("shape", [(2, 3), (128, 256), (512, 512)])
 @pytest.mark.parametrize("dtype", [torch.float32, torch.float16, torch.bfloat16])
@@ -52,7 +55,9 @@ def test_sgn__benchmark_tensor(shape, dtype):
 
     quantiles = [0.5, 0.2, 0.8]
 
-    input_tensor = torch.randn(shape, device=flag_gems.device, dtype=torch.float32).to(dtype)
+    input_tensor = torch.randn(shape, device=flag_gems.device, dtype=torch.float32).to(
+        dtype
+    )
     flat = input_tensor.view(-1)
     if flat.numel() >= 1:
         flat[0] = flat.new_zeros(1)
@@ -64,20 +69,14 @@ def test_sgn__benchmark_tensor(shape, dtype):
 
     # PyTorch reference implementation
     ms_torch, _, _ = triton.testing.do_bench(
-        lambda: torch.ops.aten.sgn_(ref_input),
-        rep=100,
-        quantiles=quantiles
+        lambda: torch.ops.aten.sgn_(ref_input), rep=100, quantiles=quantiles
     )
-
 
     # Triton implementation
     with flag_gems.use_gems():
         ms_triton, _, _ = triton.testing.do_bench(
-            lambda: gems_sgn_(act_input),
-            rep=100,
-            quantiles=quantiles
+            lambda: gems_sgn_(act_input), rep=100, quantiles=quantiles
         )
-
 
     # Calculate speedup and return result
     speedup = ms_torch / ms_triton
